kotak_neo_project/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-session>=0.8.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "neo-api-client",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.41",
    "pandas>=2.0.0",
    "requests>=2.25.1",
    "trafilatura>=1.12.2",
    "schedule>=1.2.2",
    "supabase>=0.0.3",
    "scripts>=3.0",
    "beautifulsoup4>=4.13.4",
]

[tool.uv.sources]
neo-api-client = { git = "https://github.com/Kotak-Neo/kotak-neo-api.git" }
